Splitter autentisering og autorisasjon i to pakker
