+++ server/src/main/java/no/nav/pensjon/dsf/authn/jwt/JwtUsernamePasswordAuthenticationToken.java
@@ -1,4 +1,4 @@
-package no.nav.pensjon.dsf.auth.jwt;
+package no.nav.pensjon.dsf.authn.jwt;
 
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
