+++ server/src/main/java/no/nav/pensjon/dsf/authn/ldap/LoginRequest.java
@@ -1,4 +1,4 @@
-package no.nav.pensjon.dsf.auth.ldap;
+package no.nav.pensjon.dsf.authn.ldap;
 
 class LoginRequest {
