+++ server/src/test/java/no/nav/pensjon/dsf/web/resources/ApiControllerTest.java
@@ -3,8 +3,8 @@ package no.nav.pensjon.dsf.web.resources;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.Claims;
 import no.nav.pensjon.dsf.WebServerApplication;
-import no.nav.pensjon.dsf.auth.PresysUser;
-import no.nav.pensjon.dsf.auth.jwt.JwtService;
+import no.nav.pensjon.dsf.authn.PresysUser;
+import no.nav.pensjon.dsf.authn.jwt.JwtService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
